# internal/bytealg: optimize Count/CountString in amd64